requirement.txt
flask
numpy
opencv-python
onnxruntime-gpu
faiss-cpu
insightface
Pillow
scipy
